Adjust CallersViewScopeVisitor and FlatViewScopeVisitor to handle the case of the outermost procedure scope properly.
Failing to properly attribute costs to the outermost procedure scope in a call stack profile had gone unnoticed 
before because in the past, the outermost scope was something like monitor_main. Since we had no expectation of costs
to be attributed to monitor_main, we didn't notice that the handling of outermost procedure scopes was incorrect.