Bug fix on Windows: In derived metric dialog box, when inserting a metric the inserted text is not in right positoin Cause:    Windows has very weird properties. First, losing focus means that the source widget will lose the caret position. Second, gaining the focus means that it selects all text of the widget. Therefore, when a user select a button, there's no information of the position of the caret in combo box.
Solution:
 store the information of the latest caret position in as many as SWT event possible that has impact of caret position. 
It turns out that we need to keep track of the position on Mouse-up and Key-release. I am not aware of any other event, but so far, those two events are sufficient.

Note: This happens only on XP. I don't know with Vista and Win7 ... God please give me strength.