Bug fix: recursive function in callers view doesn't show correct metric values on the 3rd  level child
Cause: there are logical flaws and programming errors in the attribution of counter (which identifies how many times the instances have been created which an indicator of recursivity).

Solution: the counter attribution needs to be divided into two steps. the initialization phase, where the counter is computed based on the number of metric "combination"; the incremental phase where the counter is computed based on the parent's counter.