Temporary dirty hack to display functions that "may" accept point-wise computation by hardcoding the dollar sign in virtual toString method. An ideal solution is to have an abstract class to extract the list, but I think it is too costly just for elegance and display.
caveat: the functions are designed to be very generic. Although we indicate that we can use them for point-wise computation, we are still able to use it for normal calculation. For instance, we can have 
   sum(1,2,3) 
which is a constant 6. We can also have 
   sum($1,$2,$3) 
which is the sum of metric 1, 2 and 3. Furthermore, we can combine both: 
   sum($1,2.4566,$2,4*10^5)

Hence having the text "sum($x1,$x2,...)" doesn't mean it disable any number as arguments.