fix feature: since aggregate function is removed from the list, and replaced with '@' sign, we need to add a mechanism so that user can choose a metric ID and insert the aggregate metric variable in the formula field.
It seems Eclipse doesn't support multiple actions in one button or one combo-box. Another solution is by having two buttons in metric assistance: one to insert metric as a "point-wise" variable, the other one is to insert metric as an "aggregate" variable.

Needs to think carefully what is the best way (the most user-friendly way) to represent two choices of actions in one feature.