Bug fix: hpcviewer causes a java exception when merging with procedure scope. theoretically, it is impossible to have more than one main entry procedures in an application. However, due to anomalies and "partial call path", hpcprof may generate more than one main entries.
Caller tree builder is not designed to handle this, and the dynamic caller scope construction raises a java exception when try to merge a scope with a procedure scope.

A hacked solution is not to merge procedure scope, and a callsite scope to have two caller (kids) instead of merging.