Partial bug fix for PFlotran. It seems if we click in a big process number, the process position is not adjusted.
For instance, if we have 8K processes and paint only 500 of them, then clicking on line 600 will cause the the painter to try get the process number of 600 out of 500. 
Hence, if we click something higher than the number of displayed processes, this causes java.nullpointer exception. (if click less than that, it luckily works unfortunately).

The fix is only carried out in the detail canvas. Since the information of processes is scattered almost everywhere we should store it in one place.