Dirty hack to avoid creating buffer twice when there is a change of depth.
On linux/GTK, if the spinner's value is change (for example from "2" to "4"), then GTK will send an event to empty the value ("") then set with the specified value ("4").
Hence, we have impression that there are two events occur in spinner instead of one. This issue will cause us to create image buffer two times: one for the empty (or zero) depth, the other one for the correct value (4).

To fix this, we can hack to consider that any empty value is illegal instead of considered as zero. Need to fix this in the future