Merging from parallel-gtk branch.
In general, multithreaded buffering works on Linux and Mac but I haven't tested on Windows yet. 
Basically, the trick here is to outline all access to graphic device interface (GDI) to be outside any paint-events (in SWT-GTK, the UI thread locks the access). Therefore, for all actions that require refresh the buffer (such as zooms, depth changes,....),  we need to create first the buffer, then fire the paint-event.

It turns out this is perhaps not the most efficient way compared to delegating the buffering within paint-event (since the system knows well when to refresh the client area). 
The trick here is to ensure that we do not create buffer more than one for the same "frame" (a frame is the trace to view). It turns out the problem is not as simple as that. For instance, on Linux/GTK, all modification on spinner widget consists of firing two events: one to empty the field, and the other one to set the value. This causes creating the image buffer twice for one change of call path depth. 
We can solve this by hacking to ignore the "empty" spinner event. But this is a hack, it is not a clean solution.

There may be other surprises, but so far it works fine on both Linux and Mac.

Note:
* I don't see big different of speed between buffering within paint-event and outside paint-event on Mac. So I think it is okay to apply this changes on Trunk.
* The multithreaded buffering works on trace view. Working for depth view is on the way.