Revert back to use single thread buffering on linux/gtk since the problem is more complicated that we thought.
It appears that some SWT events (not only paint-event) also acquires a lock to GDI resources.  This makes everything more complex and makes too many dirty hacks that will cause really spaghetti codes. 
We should merge from parallel-gtk branch to the main trunk if and only if it passes rigorous test. If the time is limited we can always merge for the revision 5.2