This is the working version with parallel buffering. In case of spinner widget, we can use SelectionListener instead of ModifyListener since the later will acquire a lock to GDI resources. This trick works fine. The caveat is that we need to ensure that spinner's value range is flexible enough for the user to select.
In case for resizing event which also requires acquiring a lock, we have to use our own Thread invoked asynchronously by UI thread. This new thread will then create the buffer in parallel assuming the lock has been released. Caveat: theoretically, there is no guarantee if the lock is already released even if we delay the buffering for 1 sec. But in practice, it works nicely :-)