Forcing the crosshair to be always inside the trace view. If the user select a new region of trace view, and the position of crosshair is not within the new region, we need to force it to be inside the cross hair.
NOTE:
This is not an elegant approach, but this helps us to avoid bugs when the cross hair is not inside the trace view.  Fixing this bug is quite tricky (which means we need to separate storage for samples inside the trace view and samples for the crosshair which makes very complex code) and a better solution is to eliminate exposing this problem to user.