Fix bug when two consecutive samples that have the same depth and call, are "painted" differently thus allowing each sample to show the depth figure.
To fix this bug "properly" we need to refactor the code. The original code treats too many special cases: the beginning, the middle and the end of samples. Furthermore, we have two different codes for painting trace view and depth view.

By refactoring the code we can have the same class (BasePaintLine) for both depth and trace view. Plus, not many special cases.

TODO: needs to handle the last sample carefully.