Add progress monitor in the status bar.
For the time being we just compute the progress work for painting the trace view, and doesn not include the time for reading XML and painting the depth view. We need to refactor the SpaceTimeData class and all painting classes to be more modular so that we can easily insert progress monitor for every work done.