Bug fix: we have to close editors that associated with the databases we want to close
Previously we close editors if the title has a certain pattern like "x-y" which is valid if the number of database is more than one. If the database opened is only one, then such pattern is not valid. We should then close editors associated with database, and check if the associated database matches with the databases we want to close.

This is just a quick fix. We should have a better and more elegant mechanism to close databases