Merge from the main trunk. Instead of merging from the summer branch to the main trunk, I want to merge from the main trunk to the summer branch first. Once it is clean and valid, we will "replace" the main trunk from whatever in the summer branch. No need to "merge" again since the merging using Eclipse can be a headache. I think it is much easier to merge with the command line.
I think there bugs in Eclipse's subclipse plugin (or in SVN Kit ?) that the merging can be so wrong. We have to be extremely CAREFUL to follow Eclipse suggestion on merging different branch. Otherwise, it is a headache.

Another solution is to use the subversive plugin. I had no problem merging with subversive (although it is also a bit tricky). Unfortunately subclipse will be the default in Eclipse.