Bug fix: when we have an "over-depth" paint, the detail paint displays the depth number. Since the summary view scan all pixels in the detail paint, it also scan the "number" and the pixel around the number.
If we don't count the pixel of the number, there are two problems:
1. we lose the correct percentage of the color of the sample, since some pixels are used to draw the number
2. In some platforms (like Cocoa), an interpolation is used if the neighbor pixels of a text are so contrast. For instance, if we display the number "8" with pixel value 0 and the surrounded pixels have the value 100, then some pixels close to the number "8" will not be 0 nor 100 (depending on the color).

Therefore, using filtering is not the right approach to compute the summary. A more valid approach is to have two images: one for painting without the numbers, and the other one is with the number. The former will be used to compute the summary, while the latter will be displayed on the view.
This cause us more memory usage, but better than giving wrong and annoying answer.

This is a hack fix, and requires code clean-up.