Bug fix. A bug reported by Gary that the second attempt of opening a plot graph in multiple database hpcviewer will cause a crash.
Cause: we have 2 kinds of data: experiment and raw metric data (also name thread level data). We have associated a window with two keys: an experiment AND a database, but we only associate a window with one key of metric data. Therefore, one opening of thread level data replace the old one.

Fix: we need to associate a window with two keys everywhere: database and experiment. this ideally causes major design change instead of a small hack. A hack might look to fix the issue, but will cause another hack in different parts and the code become very ugly. I think we should apply medium change to fix the bug and to make the code looks better.