Bug fix: handling different active sites to retrieve status bar.
In Eclipse, a status bar is hosted by either a view or an editor. Unfortunately, in some cases, the java operator "instanceof" fail to detect if an object is an instantiation of an interface or normal class. Not sure how this can occur. The only easy hack solution is to provide every possible instantiation of a site, and then retrieve its status bar.