Preparing for merging experiments.
In order to simplify the code, we should assume an merged experiment as a "normal" experiment data with "combined" database.
For this, we need to separate between a base experiment class (only contains cct), a normal experiment class (inherited from base class) and a merged class (inherited from normal class).

Note: this restructuring is not fully tested.