Preliminary draft of merging different executables:
- we adopt "weighted" ordered tree similarity approach. The nodes are ordered based on the "first" metric value, and we assume that the first metric is the most important one. By default, hpcprof outputs the SUM as the first metric.
- the "weight" of each node depends on the nodes: the class, the distance in the source code, the distance of the metric, the children and the name
- two nodes are the same if the level of "confidence" is very high: the same class, the same name, the same location, small metric distance and same children
- two nodes are similar if the level of confidence sufficiently high: the same class, small location distance, small metric distance, similar children, 
- two nodes are different if the level of confidence is low.

This level of confidence is heuristics, subjective, and very-depending of the executables. Most (if not all) tree similarity algorithm is N square or N cube, so pruning the search is highly important once we are confidence enough that two nodes are similar.

We should parameterize the level of confidence. If we know that the two executables are the same but the code has been significantly modified, or the compiler has done many transformation, we should make the "score" of confidence to be more relax.