Bug fix due to mishandling raw metrics. Since 20 March 2012, we reset the metric ID for raw metrics from 1, 1000,2,1001 to 1,2,3,4 during "</MetricTable>" parsing which is incorrect since the tag "<M>" is based on the old index which is 1,1000,2,1001, ....
Ideally resetting metric index has to be done after the parsing of XML file. Resetting the index in the middle of reading XML is always tricky.

This fix is a quick temporary solution. It doesn't work for special cases like raw metrics that the index begins at x (x is not zero).