Bug fix: when we resize hpctraceviewer window, the paint event sometimes reports that the difference of old size vs. new size is zero. This means, when the difference is zero, we don't need to refresh data. In fact, it turns out that the REAL difference is not zero, and we do have to refresh the data.
Hence, whenever we resize the window, we force SpaceTimeData class to refresh data (reading the data from the disk). 

This issue is partially Eclipse's problem. Eclipse at the moment doesn't know how to detect the last resize event in portable way. 
Worse, when resize, Eclipse doesn't propagate Mouse.Up event, thus nobody knows if we are still resizing or not.

At the moment we have to compute manually if a resize event is the last one or not. But unfortunately nothing is reliable and portable. It's highly depending on the platform.