Add basic procedure-procedure map feature. This check-in is to adapt the change of APIs in hpcdata
Recall that we have two maps:
- procedure to procedure (renaming): occurs for both hpcviewer and hpctraceviewer
- procedure to class (grouping): occurs only for hpctraceviewer to determine the color

Both map is stored in user's workspace (if installed properly will be in ~/.hpctoolkit/ otherwise in workspace directory). The location of the workspace is  determined by Eclipse, and is guaranteed to be writable. However, if the file doesn't exist, we will create it and initialize with the default.