Add base classes to map between procedure to a class. The goal is to classify static procedure with a class, and then attribute the color based on the class. By default, each procedure has its own class (thus, different color for each procedure).
We can come up with some scenarios: 
1. hash map between procedure vs. class (one-to-one)
2. glob pattern procedure vs. class (many-to-one)
3. hierarchy procedure vs. class 

For the 1st option, we have to specify the relationship for all procedure to its class. This won't be scalable.
The 2nd option is much handier. We can specify MPI* to an MPI class.
The 3rd option can handle more complicated relationship such as  MPI_Send*, MPI_*v and MPI_* 

At the moment we implement the first one since it's the easiest one, but it can be enriched to the 2nd option (the 3rd option is "future" work)