There's a bug in hpctraceviewer where the data display in trace view, depth view and call path view is not consistent. It seems either there's an issue in rendering, or problem with tracking cursor position. The problem only occurs for huge number of processes or time, when they don't fit into the display.
This patch doesn't solve the above problem, but it simplifies the code by removing processCS attribute. Currently we have two attributes to store the position of process: process and processCS. It appears that the later is used only in Call stack, and it is a relative position of process. Instead of having two attributes, it's better to have only one, and the other one can be computed easily.