Bug fix: a "ghost" selection rectangle appears when we select a zoom region, and then open another database. Hack fix: reinitialize attributes for drawing rectangle once we open a new database.
To do: clean fix. Instead of every canvas (depth, mini and detail) calls each other to update and refresh the paint, we should use "workbench event listener" instead. The idea is that each view registers a listener. And if there is a change, Eclipse fires an event to all registered views to update its content. 

A similar mechanism has been implemented when we select an option to show trace records.