Bug fix: memory leak in hpctraceviewer (part 2) The leak comes when we open another database where the viewer doesn't dispose the old color and image pair (or color table).
The reason is that it is not disposed because we have tightly-coupled relationship between views, and almost each view calls directly to each other in order to force to update the data. Therefore, when a trace view has a new data, it draws itself first, then calls depth view to redraw too. If the depth view hasn't the same data, then it will throw an exception.

The fix is just to avoid memory leak and avoid the crash. 

The ideal solution would be to restructure the code to use Eclipse's ISourceProvider mechanism to avoid calling each views directly. It is highly recommended to restructure the code for release 5.3.1