Bug fix: memory leak in hpctraceviewer (pt. 3)
The leak happens every time we open a database and create a file channel of the trace file. 
According to Oracle doc, the allocated memory due to direct byte buffer can be outside Java heap. It appears if we close the file (and the channel) we remove references to the buffer and it can be then freed.