Add an option '-j java-dir' to the viewer install script to capture the Java bin directory and add it to PATH in the viewer launch script. With this option, the path to java is fixed at install time.
Factor out the hpcviewer name in the install script so that the
install scripts for the viewer and trace viewer are the same except
for one line.  The launch scripts already do this.

Adjust how the install script resets the permissions to only add
permissions instead of setting them.

Add copyright notices to the install and launch scripts and update
them to 2012.