Bug fix from Gary: place an editor correctly in its experiment place folder
hpcviewer allows multiple database in one window, and distinguish one database to another with a split pane region. 
For instance, if an editor of database X is displayed on the left, then all editors of database X should be displayed on the left. However the current implementation only display at the current activated editor.

This patch searches which place the editor of a database is located, and then if found, activated it, and display at the same place.

Thanks for Gary to locate the bugs and fix them.