Bug fix: when hpcviewer has multiple windows and multiple database, closing some database on a window, and then exit the application cause a SWT exception since the fonts are disposed.
I believe the bug occurs when we are too early to dispose the fonts and other native resources (icons, images, ...). It seems Eclipse still try to refresh all its window children even though we are in the state of shutting down (if another window still exists).
The solution is to force to dispose resources (icons, fonts, ...) once we are really ready to exit, i.e. in postShutDown event.

A proper solution should be each window manages its own resources instead of one static resource for all windows. It requires more memory, more listener for each window, but more robust and more independent.