Bug fix: refresh the title of the view once it's activated (after being closed). Otherwise, the default name will be shown.
Apparently, there's a bug in Eclipse jface that setting the name with IViewPart doesn't do the job. Instead, setting via IViewReference does correctly.
Unfortunately, there's no way to find IViewReference from IViewPart. It seems the only way to get IViewReference is from Window's page (which is used by refreshViews method).
Hence, while this is not fixed in Indigo, we call WindowTitle.refreshViews() to force to set the view title correctly.