Porting hpctraceviewer for e4, make it works on Linux/gtk
Originally, hpctraceviewer always uses UI thread as the main thread to manage painting and to spawn multiple thread helpers. This doesn't work on e4/linux/gtk since all resource allocation (and even any getShell() methods) will attempt to acquire a lock. The impact of this strategy causes a deadlock (again only on linux/gtk).

The only proper solution is to avoid using UI thread as the main driver. All thread helpers are created by a non-UI thread, but once they attempt to use UI resource, we have to ask UI thread to do this for us, by using asyncExec (or SyncExec):

display.asyncExec( new Runnable() {
    public void run() { ...  
       getResource();
    }
});

Another issue with e4 is that all UI resources are created and defined lazily. We then have to be careful to check if the resource is created or not before use it. At the moment, hpctraceviewer works, but a bit fragile since we always assume all resources are always there.