Fix functionality to change metrics, including a derived metric.
In case of derived metrics, the impact of the change is tricky: it has to refresh the table, not only the column title.
Unfortunately, Eclipse doesn't support a reliable way to change only one column. Hence, if the table contains a derived metric, we need to refresh all the table.

This is not an optimal solution, but it does the job.