Fix and refinement to enable to modify derived metric.
Everything was designed with the assumption that a metric cannot be modified. Now we need to overhaul many classes to support metric modification:
- Experiment class: make it simpler by delegating derived metric creation into DerivedMetric class
- DerivedMetric class: support simpler creation (without needed RootScope) and modification to its formula, format and name

Bug fix: if a derived metric is created "before" the creation of callers tree (remember that callers tree is created dynamically), then "copying" (or aggregating) a derived metric from CCT to callers will result incorrect value since we "aggregate" the value from CCT to the current callers tree value.

The solution is to reset the metric values of the root before aggregating from CCT.
Change:
- RootScope: needs support to reset