Hpctraceviewer: Partially fix undo/redo on mac.
The problem was that the undo/redo operations is based on Eclipse's operation history, and the default undo will take the operation which is on the queue of the Eclipse's operation history stack. Unfortunately, the current operation is on the queue of the stack. And performing "undo" will execute the current operation instead of the previous operation.

The hack is when it is the queue of the stack is the same as the current operation, we need to do undo/redo twice. The first execution is to move the current operation from undo (or redo) stack to redo (or undo_ stack, the second execution is to perform the "real" undo/redo.

Currently the hack works well on mac. We need to check on Linux which has probably other issues.

Cosmetic update: make the Frame label easy to read, and consistent to the label on the space-time detail canvas.