Temporarily fix for handling timeout and network issues.
The quick solution is to use timeout based on the number of sleep to make it easy for developers to debug (instead of using timer). 
If the time is out, we notify the user and quit from requesting the data from the server.

There are other issues:

- What is the ideal value of timeout is not trivial. This should be configurable both from the server and client parts. 
- When the server is "up" again, it's impossible to reconnect to the server. 
- When in the middle of reading the timeout is triggered, the canvas becomes blank. Instead, we should be able to show the previous canvas (rollback to previous image).
- etc.

We may need to review the protocol for better reliability (future release).