hpctraceviewer: fixing overlapping data collection and painting.
It turns out that when the number of samples in a process line is less than 2, we skip the line entirely and do not paint it. Since the paint-threads have no knowledge how many jobs need to be done, we have to rely on the number of process lines created by data collection threads. 

Recall that "in theory" both data collection threads and paint threads know how many jobs. The problem is the paint threads don't know how many lines to be skipped.

Hence, the modification inside the paint thread:
- If the number of created process lines is equal to the number of "ideal" lines to paint, then we consider the data threads have finished the job, and the paint threads just look at the available jobs in the queue