hpctraceviewer: fix a problem when the queue is empty and the number of created timeline data is already incremented.
The reason of this issue is when we use the existing counter in SpaceTimeDataControl class, the counter is incremented while the queue is not yet added. The operation of adding to the queue and counter incrementation needs to be "atomic". Otherwise, the paint threads may enter into the situation where the queue is empty while the counter is already incremented.

Instead of using the existing counter, we just need a new atomic counter that is always decremented whether we skip a timeline (due to samples less than 2) or not. The counter is initialized with the number of "ideal" number of lines to be painted, and when it reaches to zero, it means all the jobs are done.

Minor fix:
- remove unnecessary "synchronized" attribute. I don't think we need "synchronized" within getNextTrace() method
- use the "old" way of SpaceTimeMiniCanvas. At the moment the code for the new approach is still there, but not used.

To do:
- need to fix the progress bar
- need to adopt the new approach of SpaceTimeMiniCanvas.