hpctraceviewer: bug fix in mini canvas when we filter ranks.
When a filter occurs, the visible rank range can change, it isn't from 0 to n-1 anymore. Assuming after filtering a new rank range is [r1, r2] where r1>0 and r2<n-1, then we need to avoid to allow users to create a region or to move into a region outside [r1, r2]. Recall that the mini map requires to represent ranks from 0 to n-1 so that users have an idea how the filter affects the view.

This fix is a temporary fix and we need to do more when filtering with a stride.

minor fix: clean-up the code, make the code simpler.