Major reorganization for hpctraceviewer.
I need to reorganizing hpctraceviewer to make it more modular. For instance, all classes that is designed for depth view will be placed into one package: edu.rice.cs.hpc.traceviewer.depth instead of scattered in different packages. Similarly for all views such as:
- trace view (edu.rice.cs.hpc.traceviewer.main)
- summary view (edu.rice.cs.hpc.traceviewer.summary)
- call path view and mini view (edu.rice.cs.hpc.traceviewer.misc) --> note: they are in the same view, so it makes sense to make it in the same package although their functionalities are completely independent.

In the future, we need to make all these views (or now packages) to be an independent plugin project. 

Bug fix for summary view: take into account when the image buffer is stretched to accommodate the resize of the canvas. In this case, the tooltip information needs to be adjusted to check the pixel of the canvas instead of the image buffer.

Minor enhancement hpctraceviewer: use ImageRegistry to store icons instead of storing manually. The same approach as hpcviewer. We need to take advantage of Eclipse's functionalities.