hpcviewer: bug fix when a file is included in different load modules.
hpcprof may generate the same file ID with different load module if the executable or libraries are compiled without debugging info.
For instance, in hpx/mini ghost case, the unknown file has ID = 3, but the file is "included" in different load modules:

  <LoadModuleTable>
    <LoadModule i="2" n="~unknown-load-module~"/>
    <LoadModule i="6" n="/home/johnmc/pkgs/hpx-minighost/bin/mini_ghost"/>
    <LoadModule i="164" n="/home/johnmc/pkgs/hpx-git/lib/hpx/libhpx.so.0.9.9"/>
    <LoadModule i="404" n="/home/projects/pkgs/boost-1.49.0-4.9.1/lib/libboost_thread.so.1.49.0"/>
     ...
  </LoadModuleTable>

  <FileTable>
    <File i="3" n="~unknown-file~"/>
     ...
  </FileTable>

File ID=3 is used in different load modules in the CCT:

<PF i="2" s="4" l="0" lm="2" f="3" n="4">
<PF i="4" s="166" l="0" lm="164" f="3" n="166">
<PF i="8" s="12" l="0" lm="6" f="3" n="12">
<PF i="22" s="442" l="0" lm="440" f="3" n="442">

Solution: hpcviewer needs to generate a unique file ID, and doesn't depend to file ID from hpcprof.