hpctraceviewer: bug fix on using non-midpoint painting policy Assuming a range of three samples p0, p1 and p2  where p0 is the first sample, p1 is the last sample that has the same color as p0, and p2 is the sample that has different color as p0 (and p1)  or in time line: p0 < p1 < p2
a non-midpoint policy then should have a range of p0 to p2 with p0 color.

However, the previous implementation will paint with a range of p0 to p1 with p0 color.  If p0 = p1 then it will be an empty line.

Another bug fix: to replace "?" with ".?" in the regular expression for manually procedure coloring procedures.