Share project "edu.rice.cs.hpc.traceviewer.remote" into "https://hpcviewer.googlecode.com/svn/trunk"
The "remote" version of hpctraceviewer is still in experimental mode. We don't want to release it by default. We should release separately for friendly users only.
At the moment, the current hpctraceviewer doesn't support reliable protocol. If the connection cut out, users need to reset both  the server and the viewer.

The current plugin for hpctraceviewer-remote will include automatic setup for ssh tunneling. For this, we depend on external libraries jsch and jzlib (included in external and lib folder)