hpctraceviewer: reorganize the code so that the main view can intercept exceptions (such as timeout exception) properly.
The problem with the original code is that if an exception occurs during the painting or data communication, it just returns and let other views (depth view, call stack view, summary view, ...) to continue to paint its canvas.
We should reorganize the code so that other views will NOT do any action until the main view has properly finished. To do this, only the main view can listen user operations (zoom-in, zoom-out, filter, color change, cursor position change, depth change, ...), and at the end of its operation, it will notify other views (see notifyChangeBuffer() method).

On the other hand, if an exception occurs in the middle of operations, no notification is needed, and we return to the current state (as like nothing has occurred).

Minor change: we shouldn't derive PositionOperation from TraceOperation. Both isn't strongly related, and we don't want to store position operations in undo/redo list like trace operations. It's a waste of memory.