hpctraceviewer: reorganize the code to handle more exceptions (ssh, i/o, connection  exceptions), clean-up and bug fix
Bug fix: in rev 1791, remoteConnection variable in RemoteDBOpener is declared as static (class variable), but after rev 1791, it's declared as an object variable.
The fact it was static is because the container (class RemoteDBOpener) is instantiate every time we attempt a connection whether it's the same connection or not (same connection means the same Url and the same port).

The clean way to fix this is to avoid static variables since they disable hpctraceviewer to have multiple windows like hpcviewer.  At the moment integrating hpctraceviewer and hpcviewer is not priority, so let it be different.