hpctraceviewer: minor changes by refactoring and simplifying opening local and remote database.
Instead of using a string (for local database) or an array of strings (also object for remote database) we can unify them into one class: DatabaseAccessInfo.
The class is able to tell whether the database should be local or remote. If it is a remote, it should be able to tell whether to enable SSH tunnel or not. This unification simplifies the interface for abstract db opener and trace database classes.

Other enhancements: simplify the interface for opening database in TraceDatabase class. Now we provide two interfaces to open: one for general open (by displaying open dialog box) and the other one is to open directly a local database (specified at the command line). The change removes unnecessary redundant codes in opening database. I think it would be better to delegate all the opening stuff in TraceDatabase.

Notes:
- There's a mistake (in previous commit) in SpaceTimeDataControllerLocal's constructor by assuming the input argument is a database directory. However if a user specifies (at the command line) an XML file path (instead of directory path) it will break the code. Should we check this in low level (hpcdata) ?