hpcviewer: fixing filtering mode in all views. Since we avoid to reconstruct data from scratch (creating again cct, callers view and flat view) due to filtering, we need to customize filtering for callers view and flat view. For callers view, if a procedure is excluded (filtered) then all its children are also excluded. Similarly with the flat view. However, this will cause incorrect metric value attribution since the inclusive metric value of the parent (which is the sum of inclusive metrics of its children) is not relevant anymore.
For instance, the original flat tree:

file f = 10 (100%)
- proc p1 = 8 (80%)
- proc p2 = 2 (20%)

if we filter p1, then the new flat tree:

file f = 10 (100%)   <--- INCORRECT 
- proc p2 = 2 (20%)  <--- INCORRECT 

The new tree should be instead:

file f = 2 (100%)
- proc p2 = 2 (100%)

At the moment it is not easy to have a clean solution for dynamic filtering. The easiest way is to reconstruct and recompute from scratch.