hpcviewer: finalizing filtering for CCT.
There's a bug in Eclipse that when we refresh an element in a table row, Eclipse will assume the "descendants" of the node don't change. Since filtering change the descendants, the refreshElement() will create empty rows as the descendants always exist. 
The only solution (as far as I know) is instead of using refreshElement, we pre-compute the filtering effect to not only the node, but also its children. So when we need to show the node that has filtered children, we already know the "updated" metrics, and children.

Note: this commit doesn't include filtering for bottom-up and flat tree.