hpcviewer: bug fix for activating callers view and filtering on callers view (bottom-up view).
The content of callers view is initialized only when a user "activate" the view by clicking to the tab view. If the content is not initialized, the tree is empty. However, it is possible on Eclipse that the callers view is VISIBLE but not ACTIVATED.  This occurs if we have separate tab folder between CCT view and callers view. The CCT view is by default is activated, while the callers view is visible, but is never activated.
In this case, we see an empty table in the callers view until we click the tab of the callers view.....

Another issue with caller view is that since it is initialized dynamically (and the tree also generated dynamically), the filter (which is completely a separate plugin to hpcviewer) needs to detect if it needs to create a filtered callers tree or not. This causes a strong dependency between filter plugin and hpcviewer.

One solution is to "initialize" callers tree by default even if the user doesn't activate callers view. This will cause a little bit of memory consumption but at least it is not the order of N  cube.

TODO: we still need to deal with memory issues in hpcviewer, especially if filter is enabled.